refactored streak_alerts.py
